feat: converting the deposit script into a integration test
